RagBackend/pyproject.toml
[project]
name = "asf-rag-backend"
version = "0.1.0"
description = "ASF RAG backend service"
readme = "README.md"
requires-python = ">=3.10,<3.15"


dependencies = [
    "aiofiles==24.1.0",
    "beautifulsoup4==4.13.4",
    "chardet<6",

    "camelot-py==0.11.0",
    "docx2txt>=0.9",
    "faiss-cpu>=1.12.0",
    "fastapi==0.116.1",
    "google-api-python-client==2.109.0",
    "jinja2==3.1.6",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-huggingface==0.3.1",
    "langchain-ollama==0.3.6",
    "langsmith==0.4.17",
    "layoutparser==0.3.4",
    "pdf2image==1.17.0",
    "pdfplumber==0.11.7",
    "pillow==11.3.0",
    "protobuf==6.31.1",
    "pydantic==2.11.7",
    "pyjwt==2.10.1",
    "pymysql==1.1.1",
    "pypdf2==3.0.1",
    "pytesseract==0.3.13",
    "python-docx==1.1.2",
    "python-dotenv==1.1.1",
    "python-multipart>=0.0.6",
    "redis[hiredis]>=5.0.0",
    "requests==2.32.4",
    "sentence-transformers>=5.1.0",
    "uvicorn==0.35.0",
]


[dependency-groups]
dev = [
    "black==25.1.0",
    "mypy==1.18.2",
    "pytest==8.4.2",
    "pytest-cov==7.0.0",
    "ruff==0.12.11",
    "sphinx==8.2.3; python_version >= '3.11'",
]



[tool.poetry]
package-mode = false

[tool.ruff]
line-length = 88
target-version = "py310"
extend-exclude = [".venv", "venv", "build", "dist", "__pycache__", "migrations"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]

[tool.black]
line-length = 88
target-version = ["py310"]
exclude = '''/(\.git|\.venv|venv|build|dist|__pycache__|migrations)/'''

[tool.mypy]
python_version = "3.10"
files = ["main.py", "trace_logging.py", "audit", "exception", "tests/test_tooling_smoke.py"]
ignore_missing_imports = true
check_untyped_defs = true
warn_return_any = true
warn_unused_configs = true
exclude = "(^RAG_M/|^local-KLB-files/|^backups/|^assets/)"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["trace_logging", "audit", "exception"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
